third_party/WebKit/LayoutTests/nfc/idl-NavigatorNFC.html

<!DOCTYPE html>
<script src="../resources/testharness.js"></script>
<script src="../resources/testharnessreport.js"></script>
<script>
test(function() {
  assert_true('nfc' in navigator,
              'navigator.nfc exists.');
}, 'navigator.nfc IDL test');
</script>
